moved fixtures to helper.py module hooked in conftest.py
